Begin basic abstraction
In this commit, I worked on creating the main classes used for
abstraction of the IRC protocol. I also fixed building the jar
using Maven. As part of working on the abstraction, I started
the basic networking code. I continued to follow the UML
disagrams I created to plan out my project.

To help with my understanding of network communications in Java,
I read the Oracle Java tutorial on sockets:
https://docs.oracle.com/javase/tutorial/networking/sockets/readingWriting.html

In the coming days, I want to continue to work on my abstraction -
portions of it are still incomplete. My end goal is to implement
enough of the IRC RFCs for this client to be able to be used in
most cases. I will continue to talk to my client for input.