Improve chat abstractions and implement buffer class
In this commit, I improved the code that parses messages from
the IRC server to cover more types of messages. I will continue
to work on supporting a wider variety of messages.

Following my planned UML diagram, I also implemented a Buffer
class. I did revise a few small parts of my original planning,
however, since one of the things I was planning on doing
violated principles of encapsulation.

In the next commit, I hope to work on beginning to implement
a GUI for the program.