Keep track of users in a buffer/channel
This is an important part of IRC; to keep track of who is in what
channel/buffer. When a server sends a QUIT or NICK message about a user,
it does not actually say what channels that user was in. It is up to the
client to keep track of this, and that is what this commit does.

I was originally hoping to start working on the GUI in this commit,
but the issue of keeping track of users in a channel is much more
pressing.