Begin initial work on GUI
The GUI is very simple right now, but a lot is going on behind
the scenes. I implemented threading so that the GUI and networking
code are separated. Messages are passed between the GUI and networking
code, and this commit implements the infrastructure for that
messaging.

As of right now, the chat messages for a single buffer are
displayed in the GUI. There is no ability to change what buffer
is being watched, nor see the participants in a buffer, but that
will come in later commits.