Add buffer list to GUI and implement more message passing
In this commit, I added another element to the GUI: the buffer
list. The buffer list allows the user to switch between buffers;
the message backlog is cleared upon switching and its contents
replaced with the new buffer's.

The introduction of the buffer list proved to be very interesting,
as with the introduction of it, I started getting many
ConcurrentModificationException's within my code. To solve this,
I switched to using CopyOnWriteArrayList's, which are supposed
to be able to mitigate ConcurrentModificationException's.

To test this commit, I ran the command to run the jar file
(java -jar jircc-0.0.1-SNAPSHOT.jar)