Rearrange files, add basic quit handling, and add settings GUI
In this commit, I add a settings GUI so that the user can pick
the server and the username they wish to use when connecting to
a server. This GUI was created with the help of the automatic
code generation tool WindowBuilder (an Eclipse plugin).

In the process of adding this new GUI, I rearranged the code
across a few new and different files. Although the new
arrangement deviates a bit from my original UML diagram, I
believe that these changes will make the code easier to
understand. To help with understanding, I also commented
and added javadoc comments more.

The restructured code also helped me implement a way to catch
quit handling. The program now properly sends a QUIT message to
the IRC server, instead of just closing.

I tested this commit by compiling and running the jar. I used the
new Settings GUI to connect and then tested closing the window
and making sure that the quit message was the custom one.