Add GUI elements for command input
In this commit, I add GUI elements to have command input (a text
input and a button). I use a few GridBagLayouts in order to help
position the elements in a good way.

I tested this commit by running the application and making sure
the GUI still appears correctly. I also tried resizing the app
window to make sure that the GridBagLayout works properly.