Add more user input parsing and sort listings
In this commit, I implemented the sorting of buffers and
participants within a buffer. I also implemented another command
to allow the user to part a channel. I also fixed a bug with
changing nicknames and the client not catching the nickname
change event for the user.

To test this commit, I ran the program and joined a channel with
many users in it to see if the buffers became sorted. I tried out
using the newly-implemented command. To make sure that the bug
was squashed, I did the same thing I did to discover the bug and I
also tried a few other things that were similar to the bug and
made sure that those things worked properly.